# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'UISS API documentation'
  description: 'API documentation for basic CRUD operations for specified requests. All API endpoints are implemented in JSON Schema. The data transaction must be HTTP requests sent via specified url'
  version: 1.0.0
servers:
  -
    url: 'https://api.uiss.patrickmamsery.co.tz'
paths:
...
  /api/projects:
    get:
      summary: 'Get all projects'
      description: ''
      parameters:
        -
          in: query
          name: page
          description: 'The page number to retrieve.'
          example: 1
          required: false
          schema:
            type: integer
            description: 'The page number to retrieve.'
            example: 1
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  meta:
                    current_page: 1
                    from: null
                    last_page: 1
                    path: 'https://api.uiss.patrickmamsery.co.tz/api/projects'
                    per_page: 15
                    to: null
                    total: 0
                  status: success
                  message: 'Resource retrieved successfully'
                  statusCode: 200
                properties:
                  data:
                    type: array
                    example: []
                  meta:
                    type: object
                    example:
                      current_page: 1
                      from: null
                      last_page: 1
                      path: 'https://api.uiss.patrickmamsery.co.tz/api/projects'
                      per_page: 15
                      to: null
                      total: 0
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Resource retrieved successfully'
                  statusCode:
                    type: integer
                    example: 200
      tags:
        - 'Project management'
      security: []
    post:
      summary: 'Create a new project'
      description: 'This endpoint creates a new project.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer Token'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Project management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'The title of the project.'
                  example: 'Project 1'
                description:
                  type: string
                  description: 'The description of the project.'
                  example: 'This is a project'
                category:
                  type: string
                  description: 'The category of the project.'
                  example: 'Category 1'
                owner:
                  type: string
                  description: 'The owner of the project.'
                  example: 'John Doe'
              required:
                - title
                - description
                - category
                - owner
  '/api/projects/{id}':
    get:
      summary: 'Get a single project'
      description: ''
      parameters:
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Failed to retrieve resource'
                  statusCode: 404
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Failed to retrieve resource'
                  statusCode:
                    type: integer
                    example: 404
      tags:
        - 'Project management'
      security: []
    put:
      summary: 'Update a project'
      description: 'This endpoint updates a project.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer Token'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Project management'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'The title of the project.'
                  example: 'Project 1'
                description:
                  type: string
                  description: 'The description of the project.'
                  example: 'This is a project'
                category:
                  type: string
                  description: 'The category of the project.'
                  example: 'Category 1'
                owner:
                  type: string
                  description: 'The owner of the project.'
                  example: 'John Doe'
                image:
                  type: string
                  description: 'The image of the project.'
                  example: image.png
    delete:
      summary: 'Delete a project'
      description: "This endpoint deletes a project.\n\n<aside class=\"notice\"> <strong>NOTE:</strong> This action is irreversible </aside>"
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer Token'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses: {  }
      tags:
        - 'Project management'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the project.'
        example: 9
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []